refactor(*): change file naming convention, fix flag-service DDD
