+++ back/flag-service/src/flag/FlagRepository.ts
@@ -1,0 +1,3 @@
+import { DatabaseRepository } from 'library/database/repository/DatabaseRepository';
+
+export class FlagRepository extends DatabaseRepository<any> {}
