+++ back/flag-service/src/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { FlagModule } from './module';
+import { FlagModule } from './flag/FlagModule';
 
 async function bootstrap() {
   const app = await NestFactory.create(FlagModule);
