--- back/flag-service/src/repository.ts
@@ -1,3 +1,0 @@
-import { DatabaseRepository } from 'library/database/repository/service';
-
-export class FlagRepository extends DatabaseRepository<any> {}
