+++ back/library/database/object/DatabaseObject.ts
@@ -1,0 +1,4 @@
+export class DatabaseObject {
+    _id?: string;
+    createdAt?: Date;
+}
